First commit. See message for limitations.
This only pulls in 50 posts at a time, and doesn't cycle through all your
past posts if you have more than that. So it only gives a complete backup if
you start using it before you put 50 posts and keep using it. I hope to
fix this, but since it doesn't really affect me...who knows.